fix crashing for missing plans. shift Identifiable structs to their own files
