+++ TourMate/TourMate/Frontend/Plan/Views/PlansViews/Components/IdentifiableLocation.swift
@@ -1,0 +1,16 @@
+//
+//  IdentifiableLocation.swift
+//  TourMate
+//
+//  Created by Terence Ho on 17/4/22.
+//
+
+import Foundation
+import SwiftUI
+import MapKit
+
+struct IdentifiableLocation: Identifiable {
+    var id: Int
+    var coordinate: CLLocationCoordinate2D
+    var status: PlanStatus
+}
